data-vis/requirements.txt
pandas
seaborn
numpy
matplotlib
jupyter
cassandra-driver
jupyterlab
fastf1